requirements/testing.txt
-r base.txt
pytest==4.0.1
pytest-django